add kSEQUENTIAL_ONLY as an enumerated state for the NavState to be used with socket connections.  Also change the fInputSource to be type art::InputSource and cast to art::RootInput only during the postPorocessEvent method.  This is to allow the socket connection to work for navigation
git-svn-id: http://cdcvs.fnal.gov/subversion/nusoftsvn/trunk/nutools@133 692d5501-4a7f-4c20-9605-99b1da982744

+++ EventDisplayBase/NavState.cxx
@@ -1,7 +1,7 @@
 ///
 /// \file    NavState.cxx
 /// \brief   Holds information about what action to take next.
-/// \version $Id: NavState.cxx,v 1.1.1.1 2010-12-22 16:18:52 p-nusoftart Exp $
+/// \version $Id: NavState.cxx,v 1.2 2011-08-31 20:19:58 brebel Exp $
 /// \author  messier@indiana.edu
 ///
 #include "EventDisplayBase/NavState.h"
@@ -21,6 +21,13 @@ int NavState::Which() { return gsNavState; }
 
 void NavState::Set(int which)
 {
+  // only allow the state to change if it has not been previously set
+  // to kSEQUENTIAL_ONLY, this should be made more intelligent, ie
+  // if we can only do sequential access because we are using the 
+  // socket input source, then we shouldn't be allowed to press the
+  // other buttons
+  if(gsNavState == kSEQUENTIAL_ONLY) return;
+
   gsNavState = which;
   gROOT->GetApplication()->Terminate();
 }
